Move interceptors and metrics collector packages
Move the `interceptors` package to `internal/log/interceptors` and
the `collectors` package to `server/metrics/collectors` because
they uniquely fit better there.
