+++ internal/log/interceptors/interceptors.go
@@ -1,4 +1,4 @@
-package log
+package interceptors
 
 import (
 	"context"
